- Fixed cards for api response - Changed themes
