+++ .idea/misc.xml
@@ -3,9 +3,11 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/T-Gamer/AndroidStudioProjects/MarvelHeroesActivity/MarvelHeroesApp/src/main/res/drawable/marvel_comics_thumbnail_rounded.xml" value="0.3828125" />
         <entry key="..\:/Users/T-Gamer/AndroidStudioProjects/MarvelHeroesActivity/MarvelHeroesApp/src/main/res/drawable/marvel_heroes_card_rounded.xml" value="0.3828125" />
         <entry key="..\:/Users/T-Gamer/AndroidStudioProjects/MarvelHeroesActivity/MarvelHeroesApp/src/main/res/layout/main_activity.xml" value="0.36666666666666664" />
-        <entry key="..\:/Users/T-Gamer/AndroidStudioProjects/MarvelHeroesActivity/MarvelHeroesApp/src/main/res/layout/marvel_comic_item.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/T-Gamer/AndroidStudioProjects/MarvelHeroesActivity/MarvelHeroesApp/src/main/res/layout/marvel_comic_item.xml" value="0.25" />
+        <entry key="..\:/Users/T-Gamer/AndroidStudioProjects/MarvelHeroesActivity/MarvelHeroesApp/src/main/res/layout/marvel_comics_list_fragment.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/T-Gamer/AndroidStudioProjects/MarvelHeroesActivity/MarvelHeroesApp/src/main/res/layout/marvel_heroes_activity.xml" value="0.286231884057971" />
         <entry key="..\:/Users/T-Gamer/AndroidStudioProjects/MarvelHeroesActivity/MarvelHeroesApp/src/main/res/layout/marvel_heroes_comic_item.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/T-Gamer/AndroidStudioProjects/MarvelHeroesActivity/MarvelHeroesApp/src/main/res/layout/marvel_heroes_list_fragment.xml" value="0.36666666666666664" />
