+++ MarvelHeroesApp/src/main/res/drawable/marvel_comics_thumbnail_rounded.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="#010822"/>
+
+    <corners
+        android:bottomLeftRadius="8dp"
+        android:bottomRightRadius="8dp"
+        android:topLeftRadius="42dp"
+        android:topRightRadius="42dp"/>
+</shape>
