--- MarvelHeroesApp/src/main/res/drawable/marvel_heroes_card_rounded.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#1F44AD"/>
-
-    <corners android:radius="8dp" />
-</shape>
