+++ MarvelHeroesApp/src/main/res/layout/marvel_comics_list_fragment.xml
@@ -3,8 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.lucas.marvelheroesapp.com.lucas.marvelheroes.ui.MarvelComicsListFragment">
+    android:layout_height="match_parent">
 
 <!--    <com.facebook.shimmer.ShimmerFrameLayout-->
 <!--        android:id="@+id/marvelHeroesListShimmer"-->
