+++ MarvelHeroesApp/src/main/res/values/themes.xml
@@ -2,8 +2,8 @@
     <!-- Base application theme. -->
     <style name="Theme.MarvelHeroesActivity" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
         <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/purple_500</item>
-        <item name="colorPrimaryVariant">@color/purple_700</item>
+        <item name="colorPrimary">#262A37</item>
+        <item name="colorPrimaryVariant">#262A37</item>
         <item name="colorOnPrimary">@color/white</item>
         <!-- Secondary brand color. -->
         <item name="colorSecondary">@color/teal_200</item>
